--- pages/integrations/xdd/extractions/@paperId/main.module.sass
@@ -1,16 +1,0 @@
-.feedback-flexbox
-    width: 100%
-
-    .inputs 
-        flex: 1
-        margin-right: 2em
-
-    .submit-feedback
-        flex-shrink: 0
-
-.input
-    width: 100%
-
-.popover
-    padding: .2em .5em
-    margin: 0
