+++ pages/integrations/xdd/feedback/@sourceTextID/+Page.client.module.sass
@@ -21,3 +21,21 @@
 
 .paper
   background-color: var(--secondary-background) !important
+
+.feedback-flexbox
+    width: 100%
+    margin-bottom: 2em
+
+    .inputs 
+        flex: 1
+        margin-right: 2em
+
+    .submit-feedback
+        flex-shrink: 0
+
+.input
+    width: 100%
+
+.popover
+    padding: .2em .5em
+    margin: 0
